Hack/requirements.txt
Flask==3.0.3
pdfminer.six==20231228
Werkzeug==3.0.3
Pillow==10.4.0
pytesseract==0.3.13
pdfplumber==0.11.4
docx2txt==0.8
PyMuPDF==1.24.10


